feat: make code blocks executable through Execute command
tangle now returns the tangled source code string and the language of
the code block. It is used to decide file extensions.
It also accepts a new argument add_wrapper to indicate that a wrapper
main funciton should be added to the code block if the language
necessitates it (like C, but not Python).