feat: define configuration and temp directories paths
Using the `directories` crate, this commit defines a standard location
to look for and save any application specific configuration following
the the standards of the current platform.
Also, it defines a temporaty directory location to store any needed
intermediate generated files. For *nix systems, this is a folder inside
the /tmp directory.

These locations can be overriden at runtime by setting the appropiate
environment variables (see the documentation for the `configuration`
module)