feat: execution configuration that's loadable at runtime
Refactor of the execution module into submodules and creation of a templating engine for defining simple wrappers fo source code.

Wrappers and execution code now both rely on user defined configuration that can be defined at runtime.
Execution is handled via executable scripts that receive the path to a file with the tangled source code (wrapped in the template) and must compile and run it. Any intermediate files must be placed in the same directory as the received tangled file. The scripts must also return the program's stdout, stderr and its exit code.