feat: generate default execution configuration
All files and directories in the resources/config folder of this repo
are embedded on the executable and replicated in the user's
configuration at runtime.
This allows easily populating the user's configuration dir with some
default configuration for handling code execution. Existing
configuration is not overwritten.

The build.rs script handles rebuilding the crate when a resource is
added to resources/